requirements.txt
discord.py==1.3.3
dotmap==1.3.14
